p3_poseidon2

Function add_rc_and_sbox_generic

pub fn add_rc_and_sbox_generic<F: Field, A: Algebra<F> + InjectiveMonomial<D>, const D: u64>(
    val: &mut A,
    rc: F,
)

A generic method performing the transformation:

s -> (s + rc)^D

This is a little slower than field specific implementations (particularly for packed fields) so should only be used in non performance critical places.