p3_poseidon2

Trait ExternalLayerConstructor

pub trait ExternalLayerConstructor<F, const WIDTH: usize>
where
    F: Field,
{
    // Required method
    fn new_from_constants(
        external_constants: ExternalLayerConstants<F, WIDTH>,
    ) -> Self;
}

Initialize an external layer from a set of constants.

Required Methods

fn new_from_constants( external_constants: ExternalLayerConstants<F, WIDTH>, ) -> Self

A constructor which internally will convert the supplied constants into the appropriate form for the implementation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.