p3_poseidon2

Trait MdsLightPermutation

pub trait MdsLightPermutation<T: Clone, const WIDTH: usize>: Permutation<[T; WIDTH]> { }

For the external layers we use a matrix of the form circ(2M_4, M_4, …, M_4) Where M_4 is a 4 x 4 MDS matrix. This leads to a permutation which has slightly weaker properties to MDS

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.